extract salesReport controller
